+++ src/kernel/notifications_interface.h
@@ -26,6 +26,7 @@ public:
     virtual void blockTip(SynchronizationState state, CBlockIndex& index) {}
     virtual void headerTip(SynchronizationState state, int64_t height, int64_t timestamp, bool presync) {}
     virtual void progress(const bilingual_str& title, int progress_percent, bool resume_possible) {}
+    virtual void warning(const bilingual_str& warning) {}
 };
 } // namespace kernel
