+++ src/node/kernel_notifications.h
@@ -23,6 +23,8 @@ public:
     void headerTip(SynchronizationState state, int64_t height, int64_t timestamp, bool presync) override;
 
     void progress(const bilingual_str& title, int progress_percent, bool resume_possible) override;
+
+    void warning(const bilingual_str& warning) override;
 };
 } // namespace node
